fix(WeatherPage): Autosearch for capital cities when available
Upon re-reading the requirements before submission, I realised that this is actually supposed to be limited to capital cities & displaying the weather for all of them.

The system that I'd built allows for that, but I feel like removing the additional work that was put in for searching & pinning should be showcased.

I have used a bit of an antipattern in that the homepage now dispatches a group of actions in a single batch (against redux core principles) but it now meets the requirements as specified